+++ sites/fast-component-explorer/app/components/glyph/index.ts
@@ -1,5 +1,0 @@
-import glyphSchema from "./glyph.schema";
-import Glyph from "./glyph";
-import { Icon } from "./glyph.props";
-
-export { Glyph, glyphSchema, Icon };