+++ sites/fast-component-explorer/app/utilities/text.schema.ts
@@ -1,0 +1,8 @@
+export default {
+    $schema: "http://json-schema.org/schema#",
+    title: "Text",
+    description: "A text strings schema definition.",
+    type: "string",
+    id: "text",
+    formPluginId: "text",
+};