+++ sites/site-utilities/src/index.ts
@@ -2,3 +2,4 @@ export * from "./components";
 export * from "./definitions";
 export * from "./icons";
 export * from "./react";
+export * from "./schemas";