+++ packages/react/fast-jss-manager/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.1.18](https://github.com/Microsoft/fast-dna/compare/@microsoft/fast-jss-manager@4.1.17...@microsoft/fast-jss-manager@4.1.18) (2020-06-05)
+
+**Note:** Version bump only for package @microsoft/fast-jss-manager
+
+
+
+
+
 ## [4.1.17](https://github.com/Microsoft/fast-dna/compare/@microsoft/fast-jss-manager@4.1.16...@microsoft/fast-jss-manager@4.1.17) (2020-05-18)
 
 **Note:** Version bump only for package @microsoft/fast-jss-manager