+++ sites/fast-tooling-examples/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.1](https://github.com/Microsoft/fast-dna/compare/@microsoft/fast-tooling-examples@0.1.0...@microsoft/fast-tooling-examples@0.1.1) (2020-06-05)
+
+**Note:** Version bump only for package @microsoft/fast-tooling-examples
+
+
+
+
+
 # 0.1.0 (2020-05-18)